backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==1.10.14
pytest==8.3.2
pytest-asyncio==0.23.7
pyyaml==6.0.2
